+++ admin/controls/presensi/hapus/delete.php
@@ -1,5 +1,11 @@
 <?php
 require_once $_SERVER['DOCUMENT_ROOT'] . '/azzyra-nathalyne/koneksi.php';
+
+if (!isset($_SESSION['role']) || $_SESSION['role'] != 'admin') {
+    header("Location: ../../../../logout.php");
+    exit();
+}
+
 $id = $_GET['id'];
 $sql = "DELETE FROM presensi WHERE id=$id";
 if ($conn->query($sql) === TRUE) {
